minetest-gymnasium/pyproject.toml
[project]
name = "minetest-gymnasium"
version = "0.5.0"
description = "A Gymnasium environment for minetest"
dependencies = ["gymnasium", "numpy", "pycapnp"]

[tool.ruff]
# 1. Enable flake8-bugbear (`B`) rules, in addition to the defaults.
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
]
# 2. Avoid enforcing line-length violations (`E501`)
ignore = ["E501"]

[tool.pytest.ini_options]
filterwarnings = [
  'ignore::DeprecationWarning:pygame.pkgdata.*',
  'ignore:.*PYGAME_DETECT_AVX2=1.*:RuntimeWarning',
]

[tool.setuptools]
packages = ["minetest", "minetest.proto"]

[tool.setuptools.package-data]
"minetest.proto" = ["*.capnp"]

[build-system]
requires = ["setuptools"]
